SEO pour les actus + cleanup de l'abstract
