+++ src/pages/actus/[slug].tsx
@@ -1,6 +1,5 @@
 import { Breadcrumb } from '@codegouvfr/react-dsfr/Breadcrumb';
-import { useRouter } from 'next/router';
-import { useEffect, useState } from 'react';
+import { GetStaticPaths, GetStaticProps } from 'next';
 import styled from 'styled-components';
 
 import MarkdownWrapper from '@components/MarkdownWrapper';
@@ -8,7 +7,7 @@ import SimplePage from '@components/shared/page/SimplePage';
 import Box from '@components/ui/Box';
 import Heading from '@components/ui/Heading';
 import Text from '@components/ui/Text';
-import { getArticle } from '@data/contents';
+import { articles, getArticle } from '@data/contents';
 import { Article } from 'src/types/Article';
 
 const ArticleContentWrapper = styled(Box)`
@@ -30,29 +29,13 @@ const ThemeTagItem = styled.li`
   line-height: 1.8em !important;
 `;
 
-const ActualitePage = () => {
-  const router = useRouter();
-  const [article, setArticle] = useState<Article | null>(null);
-
-  useEffect(() => {
-    if (!router.query.slug) {
-      return;
-    }
-
-    const article = getArticle(router.query.slug as string);
-    if (article) {
-      setArticle(article);
-    } else {
-      router.push('/actus');
-    }
-  }, [router.query]);
-
+const ActualitePage: React.FC<{ article: Article }> = ({ article }) => {
   const lines = article?.content?.split('\n') ?? [];
   const title = lines.slice(0, 1)?.[0]?.substring(1); // remove the hashtag
   const content = lines.slice(1).join('\n');
 
   return (
-    <SimplePage currentPage="/actus" title={title}>
+    <SimplePage currentPage="/actus" title={title} description={article.abstract}>
       <Box className="fr-container fr-mb-n2w fr-mb-md-n4w">
         <Breadcrumb
           currentPageLabel={title}
@@ -77,7 +60,7 @@ const ActualitePage = () => {
           </Heading>
 
           <Text className="fr-icon-arrow-right-line" my="2w">
-            Publié le {article?.publishedDate.toLocaleDateString('fr-FR')}
+            Publié le {new Date(article?.publishedDate).toLocaleDateString('fr-FR')}
           </Text>
 
           <ul className="fr-tags-group">
@@ -95,6 +78,37 @@ const ActualitePage = () => {
       </ArticleContentWrapper>
     </SimplePage>
   );
+};
+
+export const getStaticProps: GetStaticProps = async ({ params }) => {
+  const slug = params?.slug as string;
+  const article = getArticle(slug);
+
+  if (!article) {
+    return {
+      redirect: {
+        destination: `/actus?notify=error:${encodeURIComponent("Désolé, cette actualité n'existe pas.")}`,
+        permanent: false,
+      },
+    };
+  }
+
+  return {
+    props: {
+      article: JSON.parse(JSON.stringify(article)) as typeof article,
+    },
+  };
+};
+
+export const getStaticPaths: GetStaticPaths = async () => {
+  const paths = articles.map(({ slug }) => ({
+    params: { slug },
+  }));
+
+  return {
+    paths,
+    fallback: 'blocking',
+  };
 };
 
 export default ActualitePage;
